intel: Update to reflect that we have bootable images created
With the corresponding changes merged to meta-agl we create wic images,
and these can be uncompressed and booted directly.  We continue to
reference mkefi-agl.sh as it used in some cases.  We no longer have
hddimg files so all referneces to those are removed.  Borrow the section
from Raspberry Pi about using dd to write an image.

Bug-AGL: SPEC-963
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/#/c/12159/
Signed-off-by: Tom Rini <trini@konsulko.com>